Simplify validation
Introduce a utility function which gets rid of the duplication when
validating objects with the `class-validator` library. Also adapt the
error controller to serialize the details of those errors in the
response as JSON instead of just sending the string message. This is
more machine-readable in case the frontend needs to react to specific
errors.

Instead of:

    {
      "code": 400,
      "message": "An instance of User has failed the validation:\n - property pseudonym has failed the following constraints: isString \n,..."
    }

The response body is similar to this:

    {
      "code": 400,
      "message": "A validation error occurred",
      "errors": [
          {
              "children": [],
              "constraints": {
                  "isString": "pseudonym must be a string"
              },
              "property": "pseudonym",
              "target": {...}
          },
          ...
      ]
    }