Improve development & production environments
The backend is no longer executed through `ts-node` which is slow. `tsc`
is used to watch and compile the TypeScript source code, and `nodemon`
to watch and execute the copiled JavaScript code in the new `dist`
directory.

Additionally, now that migration scripts have been put in place, the
database is no longer dropped and recreated every time the application
starts. Convenience scripts to migrate and roll back database migrations
have been added to `package.json`.

The previous `Dockerfile` has been moved to `Dockerfile.dev`, while the
new `Dockerfile` can be used to make a production image.